% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{cite_phines}
\alias{cite_phines}
\title{Translate values to Phenological States}
\usage{
cite_phines(palette, value)
}
\arguments{
\item{palette}{Name of the palette.}

\item{value}{A numeric value between 0 and 1.}
}
\value{
A string representing the biological stage and scientific code.
}
\description{
Translate values to Phenological States
}
\examples{
cite_phines("palay_harvest", 0.8)
cite_phines("forest_transition", 0.2)
}
