% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{show_phines}
\alias{show_phines}
\title{Visual Preview of Palettes}
\usage{
show_phines(name)
}
\arguments{
\item{name}{Palette name.}
}
\description{
Visual Preview of Palettes
}
\examples{
show_phines("coral_bleach")
}
