% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{phines_metadata}
\alias{phines_metadata}
\title{Philippine Phenological Palette Metadata}
\format{
An object of class \code{list} of length 5.
}
\usage{
phines_metadata
}
\value{
A nested \code{list} where each element represents a palette containing 
vectors of colors, biological stages, and scientific codes.
}
\description{
Internal list containing hex codes and scientific stages (BBCH, MSI, and ST-Mod).
}
\keyword{datasets}
