% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_fill_phines}
\alias{scale_fill_phines}
\alias{scale_color_phines}
\title{ggplot2 Scales for palettephines}
\usage{
scale_fill_phines(palette = "palay_harvest", discrete = TRUE, ...)

scale_color_phines(palette = "palay_harvest", discrete = TRUE, ...)
}
\arguments{
\item{palette}{Palette name.}

\item{discrete}{Logical, if TRUE returns a discrete scale.}

\item{...}{Passed to ggplot2 scale functions.}
}
\value{
A \code{ggplot2} scale object (class \code{ScaleDiscrete} or 
\code{ScaleContinuous}) to be added to a ggplot object. This maps 
data values to the phenological colors of the Philippines.
}
\description{
ggplot2 Scales for palettephines
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg, fill = factor(cyl))) +
   geom_point(shape = 21, size = 3) +
   scale_fill_phines("palay_harvest")
}
