% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{show_phines}
\alias{show_phines}
\title{Visual Preview of Palettes}
\usage{
show_phines(name)
}
\arguments{
\item{name}{Palette name.}
}
\value{
No return value, called for side effects. This function generates 
a plot in the active graphics device showing the color blocks of the palette.
}
\description{
Visual Preview of Palettes
}
\examples{
# Use oldpar to respect user's graphical settings
oldpar <- par(no.readonly = TRUE)
show_phines("coral_bleach")
par(oldpar)
}
